[conventions] Add section on concept naming
Concepts should be named in CamelCase as either nouns or adjectives.
This is the only convention for concepts.  Because we've used
lowercase-separated-by-underscores for classes, we should have no
collisions, and it should be immediately apparent whether an identifier
is a concept or a type name.

We also want to make sure we avoid template declarations like

    template <Integral IntegralT> …

which doesn't add anything but typing.  This requires a modification to
the template parameter section.

License: CC-BY4.0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>